+++ MainProject/TankAttributes/IDamageable.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.Design.Serialization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -8,6 +9,6 @@ namespace MainProject.TankAttributes
 {
     internal interface IDamageable
     {
-        void Damaged(int damage);
+        void Damaged(float damage);
     }
 }
